refactor: Move AbstractDistributedLock to the lock module
